% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rand_kdpi_cold.R
\name{rand_kdpi_cold}
\alias{rand_kdpi_cold}
\title{Random draw from joint distribution of KDPI and cold ischemia time}
\usage{
rand_kdpi_cold(seed = 11111, n, france = FALSE)
}
\arguments{
\item{seed}{A seed number}

\item{n}{Number of draws needed}

\item{france}{KDPI distribution from France?}
}
\value{
two numerical vectors
}
\description{
Random draw from joint distribution of KDPI and cold ischemia time
}
\examples{
values <- rand_kdpi_cold(seed=12345, n=20, france=FALSE)
}
