% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_simulation.R
\name{run_simulation}
\alias{run_simulation}
\title{Run the simulation}
\usage{
run_simulation(
  seed,
  data = sample_df,
  ddtx_coef = ddtx_coef_w,
  ddtx_p = ddtx_p_w,
  ldtx_coef = ldtx_coef_g,
  ldtx_gamma = ldtx_gamma_g,
  mort_coef = mort_coef_w,
  mort_shape = mort_p_w,
  remove_coef = remove_coef_w,
  remove_p = remove_p_w,
  ddtx_rate = 0,
  gl30_coef = gl30_coef_ml_noreg,
  dgf_coef = dgf_coef_ml_noreg,
  mort30_coef = mort30_coef_ml_noreg,
  gl_coef = gl_coef_g_noreg,
  gl_shape = gl_gamma_g_noreg,
  gs_mort_coef = gs_mort_coef_g_noreg,
  gs_mort_shape = gs_mort_gamma_g_noreg,
  dial_mort_coef = dial_mort_coef_w_noreg,
  dial_mort_shape = dial_mort_lnp_w_noreg,
  gl_mort_coef = gl_mort_coef_logit_noreg,
  france = FALSE
)
}
\arguments{
\item{france}{}
}
\description{
Run the simulation
}
\examples{
results <- run.simulation2(seed = 11111)
}
