% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_simulation.R
\name{run_simulation}
\alias{run_simulation}
\title{Run the simulation}
\usage{
run_simulation(
  seed,
  data = sample_df,
  ddtx_coef,
  ddtx_p,
  ldtx_coef,
  ldtx_gamma,
  mort_coef,
  mort_shape,
  remove_coef,
  remove_p,
  ddtx_rate = 0,
  france = FALSE,
  gl30_coef,
  dgf_coef,
  mort30_coef,
  gl_coef,
  gl_shape,
  gs_mort_coef,
  gs_mort_shape,
  dial_mort_coef,
  dial_mort_shape,
  gl_mort_coef
)
}
\arguments{
\item{seed}{random seed}

\item{data}{data of simulated individuals}

\item{ddtx_coef}{deceased donor transplant coefficients}

\item{ddtx_p}{deceased donor transplant shape parameter}

\item{ldtx_coef}{living donor transplant coefficients}

\item{ldtx_gamma}{living donor transplant shape parameter}

\item{mort_coef}{waitlist mortality coefficients}

\item{mort_shape}{waitlist mortality shape parameter}

\item{remove_coef}{waitlist removal coefficients}

\item{remove_p}{waitlist removal shape parameter}

\item{ddtx_rate}{increased rate of transplantation?}

\item{france}{use KDPI distribution from France?}

\item{gl30_coef}{30-Day Multinomial Logit: graft loss coefficients}

\item{dgf_coef}{30-Day Multinomial Logit: delayed graft function coefficients}

\item{mort30_coef}{30-Day Multinomial Logit: mortality coefficients}

\item{gl_coef}{graft loss coefficients}

\item{gl_shape}{graft loss shape parameter}

\item{gs_mort_coef}{death with function coefficients}

\item{gs_mort_shape}{death with function shape parameter}

\item{dial_mort_coef}{death after graft loss coefficients}

\item{dial_mort_shape}{death after graft loss shape parameter}

\item{gl_mort_coef}{death same day as graft loss coefficients}
}
\value{
data frame with results for each individual
}
\description{
Run the simulation
}
\examples{
results <- run.simulation2(seed = 11111)
}
